backend/requirements.txt
Django==6.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==21.2.0
uvicorn==0.27.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
redis==5.0.1
channels==4.0.0
daphne==4.0.0
Pillow==10.1.0
requests==2.31.0
celery==5.3.4
django-filter==23.5
djangorestframework-filters==1.0.0.dev0
drf-nested-routers==0.93.4
django-extensions==3.2.3
black==23.12.1
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
